requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-teddynote
PyPDF2
faiss-cpu
python-dotenv
google-generativeai
pillow
tiktoken
pymupdf

